Bug6

Description

Exporting records.

Documentation

data A

This record is exported without its field

Constructors

A Int

data B

.. with its field, but the field is named separately in the export list (the field isn't documented separately since it is already documented here)

Constructors

B

Fields

b :: Int

data C

.. with fields names as subordinate names in the export

Constructors

C

Fields

c1 :: Int

c2 :: Int

data D

.. with only some of the fields exported (we can't handle this one - how do we render the declaration?)

Constructors

D Int Int

newtype E

a newtype with a field

Constructors

E Int